refactor: InputDataと関連するクラスを別ファイルに切り出す
